+++ EUExCreditCardRec/EUExCreditCardRec/sourceCode/CardIO/CardIO.h
@@ -1,0 +1,15 @@
+//
+//  CardIO.h
+//  Version 5.2.2
+//
+//  See the file "LICENSE.md" for the full license governing this code.
+//
+
+// All-in-one header file for card.io sdk.
+#import "CardIOCreditCardInfo.h"
+#import "CardIODetectionMode.h"
+#import "CardIOView.h"
+#import "CardIOViewDelegate.h"
+#import "CardIOPaymentViewController.h"
+#import "CardIOPaymentViewControllerDelegate.h"
+#import "CardIOUtilities.h"
