--- EUExCreditCardRec/EUExCreditCardRec/sourceCode/PayPalMobile/CardIODetectionMode.h
@@ -1,18 +1,0 @@
-//
-//  CardIODetectionMode.h
-//  Version 4.0.2
-//
-//  See the file "LICENSE.md" for the full license governing this code.
-//
-//
-
-#ifndef icc_CardIODetectionMode_h
-#define icc_CardIODetectionMode_h
-
-typedef NS_ENUM(NSInteger, CardIODetectionMode) {
-  CardIODetectionModeCardImageAndNumber = 0,
-  CardIODetectionModeCardImageOnly,
-  CardIODetectionModeAutomatic
-};
-
-#endif
