+++ EUExCreditCardRec/EUExCreditCardRec/sourceCode/PayPalMobile/PayPalConfiguration.h
@@ -1,7 +1,7 @@
 //
 //  PayPalConfiguration.h
 //
-//  Version 2.8.2
+//  Version 2.12.7
 //
 //  Copyright (c) 2014, PayPal
 //  All rights reserved.
@@ -24,6 +24,8 @@ typedef NS_ENUM(NSInteger, PayPalShippingAddressOption) {
 /// @see forceDefaultsInSandbox
 @property(nonatomic, copy, readwrite) NSString *defaultUserEmail;
 /// Optional default user phone country code used in the PayPal login view.
+/// The input to this method is expected to be a digit string.
+/// For example: `@"1"` for North America, `@"44"` for UK
 /// Will be overridden by phone country code used in most recent PayPal login.
 /// @see forceDefaultsInSandbox
 @property(nonatomic, copy, readwrite) NSString *defaultUserPhoneCountryCode;
@@ -85,8 +87,19 @@ typedef NS_ENUM(NSInteger, PayPalShippingAddressOption) {
 /// E.g., specifying "en" on a device set to "English" and "United Kingdom" will result in "en_GB".
 ///
 /// These localizations are currently included:
-/// ar,da,de,en,en_AU,en_GB,en_SE,es,es_MX,fr,he,it,ja,ko,ms,nb,nl,pl,pt,pt_BR,ru,sv,th,tr,zh-Hans,zh-Hant,zh-Hant_TW.
+/// ar,da,de,en,en_AU,en_GB,es,es_MX,fr,he,it,ja,ko,ms,nb,nl,pl,pt,pt_BR,ru,sv,th,tr,zh-Hans,zh-Hant,zh-Hant_TW.
 @property(nonatomic, copy, readwrite) NSString *languageOrLocale;
+
+/// If set to YES, then all displayed payment amounts will include a currency code (e.g., "USD")
+/// in addition to the standard currency symbol (e.g., "$").
+///
+/// Even if set to NO, currency codes will still be displayed in the case where a user's PayPal account
+/// includes any funding source whose currency differs from that of the current payment.
+/// For example, if the current payment is in US Dollars but the user's PayPal account includes a
+/// credit card in Canadian Dollars, then all displayed amounts will be labeled as either USD or CAD.
+///
+/// Defaults to NO.
+@property(nonatomic, assign, readwrite) BOOL alwaysDisplayCurrencyCodes;
 
 /// Normally, the SDK blurs the screen when the app is backgrounded,
 /// to obscure credit card or PayPal account details in the iOS-saved screenshot.
