+++ EUExCreditCardRec/EUExCreditCardRec/sourceCode/PayPalMobile/PayPalMobile.h
@@ -1,7 +1,7 @@
 //
 //  PayPalMobile.h
 //
-//  Version 2.8.2
+//  Version 2.12.7
 //
 //  Copyright (c) 2014, PayPal
 //  All rights reserved.
@@ -54,7 +54,7 @@ extern NSString *const PayPalEnvironmentNoNetwork;
 + (void)preconnectWithEnvironment:(NSString *)environment;
 
 /// Once a user has consented to future payments, when the user subsequently initiates a PayPal payment
-/// from their device to be completed by your server, PayPal uses a Correlation ID to verify that the
+/// from their device to be completed by your server, PayPal uses a Client Metadata ID to verify that the
 /// payment is originating from a valid, user-consented device+application.
 /// This helps reduce fraud and decrease declines.
 /// This method MUST be called prior to initiating a pre-consented payment (a "future payment") from a mobile device.
@@ -63,7 +63,7 @@ extern NSString *const PayPalEnvironmentNoNetwork;
 /// @return clientMetadataID Your server will send this to PayPal in a 'PayPal-Client-Metadata-Id' header.
 + (NSString *)clientMetadataID;
 
-/// Method deprecated use clientMetadataID instead
+/// Method deprecated. Use clientMetadataID instead
 + (NSString *)applicationCorrelationIDForEnvironment:(NSString *)environment __attribute__((deprecated("Use clientMetadataID instead.")));
 
 /// Delete all previously remembered user data (credit card, email, phone, PayPal account),
