+++ EUExCreditCardRec/EUExCreditCardRec/sourceCode/PayPalMobile/PayPalOAuthScopes.h
@@ -1,7 +1,7 @@
 //
 //  PayPalOAuthScopes.h
 //
-//  Version 2.8.2
+//  Version 2.12.7
 //
 //  Copyright (c) 2014, PayPal
 //  All rights reserved.
@@ -11,17 +11,17 @@
 // @see https://developer.paypal.com/docs/integration/direct/identity/attributes/ for more details
 
 /// Authorize charges for future purchases paid for with PayPal.
-NSString *const kPayPalOAuth2ScopeFuturePayments;
+extern NSString *const kPayPalOAuth2ScopeFuturePayments;
 /// Share basic account information.
-NSString *const kPayPalOAuth2ScopeProfile;
+extern NSString *const kPayPalOAuth2ScopeProfile;
 /// Basic Authentication.
-NSString *const kPayPalOAuth2ScopeOpenId;
+extern NSString *const kPayPalOAuth2ScopeOpenId;
 /// Share your personal and account information.
-NSString *const kPayPalOAuth2ScopePayPalAttributes;
+extern NSString *const kPayPalOAuth2ScopePayPalAttributes;
 /// Share your email address.
-NSString *const kPayPalOAuth2ScopeEmail;
+extern NSString *const kPayPalOAuth2ScopeEmail;
 /// Share your account address.
-NSString *const kPayPalOAuth2ScopeAddress;
+extern NSString *const kPayPalOAuth2ScopeAddress;
 /// Share your phone number.
-NSString *const kPayPalOAuth2ScopePhone;
+extern NSString *const kPayPalOAuth2ScopePhone;
