+++ EUExCreditCardRec/EUExCreditCardRec/sourceCode/PayPalMobile/PayPalPaymentViewController.h
@@ -1,7 +1,7 @@
 //
 //  PayPalPaymentViewController.h
 //
-//  Version 2.8.2
+//  Version 2.12.7
 //
 //  Copyright (c) 2014, PayPal
 //  All rights reserved.
