--- EUExCreditCardRec/EUExCreditCardRec/sourceCode/cardioLib/CardIO.h
@@ -1,10 +1,0 @@
-//
-//  CardIO.h
-//  Copyright (c) 2011-2013 PayPal. All rights reserved.
-//
-
-// All-in-one header file for card.io sdk.
-
-#import "CardIOPaymentViewController.h"
-#import "CardIOPaymentViewControllerDelegate.h"
-#import "CardIOCreditCardInfo.h"
